Explicit table name for counseling
In debugging the docker container on dev, the implicit naming strategy does not seem to work, and we get errors indicating that the table "coach.Counseling" does not exist. It's not clear why this is only happening in a containerized dev instance and nowhere else, but let's see if explicit naming resolves it.